lots 'o' stuff
Fixed a bunch of annoying Cordova plugin problems, fixed some bugs and
got everything to load properly. Modified some loading code, and also
put in a bunch more log messages in preparation for the debug feature.
Take a look at the console in ripple to see what is happening when. Also
added windows8 as a target for debugging